requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.3
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-jwt==1.11.0
drf-yasg==1.21.8
idna==3.10
inflection==0.5.1
packaging==24.2
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyJWT==1.7.1
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
